requirements.txt
Flask
newsapi-python
newspaper3k
langchain
langchain_groq
